+++ addOns/selenium/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Changed
+- Now using 2.10 logging infrastructure (Log4j 2.x).
 - Update links to zaproxy repo.
 - Maintenance changes.